Store user in DB upon sign in, load from DB upon auth state changed
The handleSignInSuccessWithAuthResult callback gives us lots of rich data and indicates a difinitive login. However, it does not fire upon returning to the page with a session already established; we must listen for that in onAuthStateChanged. Therefore, we save the rich data to the user's profile in the former callback and load it up in the latter.