Don't actually need Rebase
It looks like it only gives us some data binding magic (which we don't use in hooks anyway) and perhaps the "syncDoc" method (which doesn't seem all that reactive -- prefer to update a document in response to input, listen to changes to collections, update and components accordingly). Also trusting the Google SDK to be more performant, reliable, etc.