Fix endless loop when some props were empty arrays
This is a two-sided bug: on one hand, the 'allPropsLoaded()' method considers an empty array as false, so that means the actual component will never be rendered unless some record exists in the array.

On the other hand, passing an empty array ([]) to the 'useCoachAssignments()' hook is interpreted as a new object every time, so the component effect is being rendered endlessly.