Fix Facebook warnings (again)
The only downside to the current implementation is that it doesn't allow for Advanced Tracking upon login (rather, only when a user revisits the site after being already logged in); unfortunately, I can't think of an easy way around this since the Pixel SDK doesn't have any kind of un-init, other than forcing a hard page reloag or toying with two unique Pixels on the page (one for unauthenticated, one for authenticated), which both seem sloppy. Revisit if needed.