Reduce deep state passsing w/ light-weight event pub/sub
Because the Action and Sentiment elements don't have a parent/child relationship, the cleanest way to communicate between them is to use a central event emitting mechanism. Redux usually plays this role, but since we're not using it (yet?), we use a simple alternative here. Note that RxJS was looked into, but doesn't handle centrality (subscriptions must accept a subject object rather than a simple string, meaning we would have to pass ssubjects around in a provider, negating any bennefit).